Adding changes that allow calculating transmission or reflection off a layer or a medium bound.
Currently there is a bug when we transmit through a layer that happens
to be at the top of the medium or bottom of the medium, since it makes
the currentLayer null.
Need to switch the order of the checks to see if the photon hit the medium
before it checks to see if a layer boundary was hit/crossed.