Fixed bug with assertions failing in photon.cpp
Assertions of ALIVE and hit_medium_boundary were failing
due to the reset() call when running the 'runAcoustoOptics()'.
Reset was bringing the state of the photon back to ALIVE after
it had died, but we do not want this to happen when the iterations
of this object are 1.

Now the exit-aperture data is only being written for the first time
step of the AO portion.  Needs debugging.