Fixed bug with freeing photon objects causing segfault.
The issue arised when the loop in 'runAcoustoOptics()' looked
for the count of 'Logger::getInstance()->numDetectedPhotons()'.
This would always be zero if the seed file had previously been
generated (i.e. 'generateSeeds()' was not run beforehand) and was
used (which is the end goal since we don't want to have to generate
these seeds everytime).  Therefore the number of detected photons
was zero and the loop would never create the photon objects on the
heap, which meant I was trying to free unallocated objects.  Segfault.

This now simply reads in the seed file from stream and based on that
size (which is the same as the Logger would have contained) we form
the bounds of the for loop.